Refactor auth

+++ src/services/services.service.ts
@@ -1,94 +1,116 @@
 import {
-  BadRequestException,
   Injectable,
-  InternalServerErrorException,
-  Logger,
   NotFoundException,
+  InternalServerErrorException,
+  BadRequestException,
 } from '@nestjs/common';
 import { CreateServiceDto } from './dto/create-service.dto';
 import { UpdateServiceDto } from './dto/update-service.dto';
 import { DataSource, In, Repository } from 'typeorm';
 import { InjectRepository } from '@nestjs/typeorm';
 import { PaginationDto } from 'src/common/dto/pagination.dto';
-import { User } from 'src/auth/entities/user.entity';
-import { ServiceImage, Service } from './entities';
+import { Service } from './entities';
 import { Staff } from 'src/staff/entities/staff.entity';
+import {
+  CreateServiceResponse,
+  DeleteServiceResponse,
+  GetServiceResponse,
+  GetServicesResponse,
+  UpdateServiceResponse,
+} from './interfaces';
 
 @Injectable()
 export class ServicesService {
-  private readonly logger = new Logger('ServicesService');
-
   constructor(
     @InjectRepository(Service)
     private readonly serviceRepository: Repository<Service>,
-    @InjectRepository(ServiceImage)
-    private readonly serviceImageRepository: Repository<ServiceImage>,
     @InjectRepository(Staff)
     private staffRepository: Repository<Staff>,
 
     private readonly dataSource: DataSource,
   ) {}
 
-  async create(createServiceDto: CreateServiceDto, user: User) {
+  async create(
+    createServiceDto: CreateServiceDto,
+  ): Promise<CreateServiceResponse> {
     try {
-      const { images = [], staff: staffIds, ...rest } = createServiceDto;
+      const { staff: staffIds, ...rest } = createServiceDto;
 
       const staff = await this.staffRepository.findBy({
         id: In(staffIds),
       });
+      if (!staff) throw new BadRequestException(`Provided staff  not found.`);
 
       const service = this.serviceRepository.create({
-        user,
-        images: images.map((img) =>
-          this.serviceImageRepository.create({ url: img }),
-        ),
         staff,
         ...rest,
       });
       await this.serviceRepository.save(service);
       return service;
     } catch (error) {
-      this.handleDBExceptions(error);
+      console.error(error);
+      throw new InternalServerErrorException(
+        'An unexpected error occurred while creating service. Please try again later.',
+        'CREATE_SERVICE_FAILED',
+      );
     }
   }
 
-  async findAll(paginationDto: PaginationDto) {
-    const { limit = 10, offset = 0 } = paginationDto;
-    const services = await this.serviceRepository.find({
-      take: limit,
-      skip: offset,
-      relations: {
-        images: true,
-        staff: true,
-      },
-      order: {
-        name: 'ASC',
-      },
-    });
+  async findAll(paginationDto: PaginationDto): Promise<GetServicesResponse> {
+    try {
+      const { limit = 10, offset = 0 } = paginationDto;
+      const services = await this.serviceRepository.find({
+        take: limit,
+        skip: offset,
+        relations: {
+          staff: true,
+        },
+        order: {
+          name: 'ASC',
+        },
+      });
 
-    return services;
+      return services;
+    } catch (error) {
+      console.error(error);
+      throw new InternalServerErrorException(
+        'An unexpected error occurred while finding services. Please try again later.',
+        'GET_SERVICES_FAILED',
+      );
+    }
   }
 
-  async findOne(id: string) {
-    const service = await this.serviceRepository.findOne({
-      where: {
-        id,
-      },
-      relations: {
-        images: true,
-        staff: true,
-      },
-    });
-    if (!service) throw new NotFoundException();
+  async findOne(id: string): Promise<GetServiceResponse> {
+    try {
+      const service = await this.serviceRepository.findOne({
+        where: {
+          id,
+        },
+        relations: {
+          staff: true,
+        },
+      });
+      if (!service) throw new NotFoundException();
 
-    return service;
+      return service;
+    } catch (error) {
+      console.error(error);
+      throw new InternalServerErrorException(
+        'An unexpected error occurred while finding services. Please try again later.',
+        'GET_SERVICE_FAILED',
+      );
+    }
   }
 
-  async update(id: string, updateServiceDto: UpdateServiceDto, user: User) {
-    const { images, staff, ...rest } = updateServiceDto;
+  async update(
+    id: string,
+    updateServiceDto: UpdateServiceDto,
+  ): Promise<UpdateServiceResponse> {
+    const { staff, ...rest } = updateServiceDto;
     const dbStaff = await this.staffRepository.findBy({
       id: In(staff),
     });
+    if (!dbStaff) throw new BadRequestException('Provided staff not found');
     const service = await this.serviceRepository.preload({
       id,
       staff: dbStaff,
@@ -102,14 +124,6 @@ export class ServicesService {
     await queryRunner.startTransaction();
 
     try {
-      if (images) {
-        await queryRunner.manager.delete(ServiceImage, { service: { id: id } });
-        service.images = images.map((img) =>
-          this.serviceImageRepository.create({ url: img }),
-        );
-      }
-
-      service.user = user;
       await queryRunner.manager.save(service);
       await queryRunner.commitTransaction();
       await queryRunner.release();
@@ -118,24 +132,29 @@ export class ServicesService {
     } catch (error) {
       await queryRunner.rollbackTransaction();
       await queryRunner.release();
-
-      this.handleDBExceptions(error);
+      console.error(error);
+      throw new InternalServerErrorException(
+        'An unexpected error occurred while updating service. Please try again later.',
+        'UPDATE_SERVICE_FAILED',
+      );
     }
   }
 
-  async remove(id: string) {
-    const service = await this.findOne(id);
-    await this.serviceRepository.remove(service);
-    return service;
-  }
+  async remove(id: string): Promise<DeleteServiceResponse> {
+    try {
+      const service = await this.serviceRepository.findOneBy({ id });
+      if (!service)
+        throw new NotFoundException(`Service with id ${id} not found`);
 
-  // Handle Errors
-  private handleDBExceptions(error: any) {
-    if (error.code === '23505') throw new BadRequestException(error.detail);
+      await this.serviceRepository.remove(service);
 
-    this.logger.error(error);
-    throw new InternalServerErrorException(
-      'Unexpected error, check server logs.',
-    );
+      return service;
+    } catch (error) {
+      console.error(error);
+      throw new InternalServerErrorException(
+        'An unexpected error occurred while deleting service. Please try again later.',
+        'DELETE_SERVICE_FAILED',
+      );
+    }
   }
 }
